pyproject.toml
[tool.poetry]
name = "zohoinfobot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
aiogram = {extras = ["all"], version = "^2.25.1"}
aiofiles = "^23.1.0"
django = "^5.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
httpx = "^0.25.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
